requirements.txt
numpy
pillow
scipy
matplotlib
pycryptodome
fastapi
uvicorn[standard]
boto3
zlib
pynacl
cryptography
hmac
google-cloud-storage
